+++ src/stack-unix/tcpv4_socket.mli
@@ -16,7 +16,6 @@
 
 include Mirage_protocols.TCP
   with type ipaddr = Ipaddr.V4.t
-   and type ipinput = unit Lwt.t
    and type flow = Lwt_unix.file_descr
    and type error = [ Mirage_protocols.Tcp.error | `Exn of exn ]
    and type write_error = [ Mirage_protocols.Tcp.write_error | `Exn of exn ]
@@ -24,3 +23,5 @@ include Mirage_protocols.TCP
 val connect : Ipaddr.V4.Prefix.t -> t Lwt.t
 
 val disconnect : t -> unit Lwt.t
+
+val set_switched_off : t -> unit Lwt.t -> unit